Altera exercícios iniciais

+++ exercicio01.py
@@ -1,5 +1,6 @@
-numero_1 = int(input('Digite o primeiro número: '))
-numero_2 = int(input('Digite o segundo número: '))
+numero_1 = float(input('Digite o primeiro número: '))
+numero_2 = float(input('Digite o segundo número: '))
 
 produto = numero_1 * numero_2
+
 print(f'O produto dos números é {produto}')
